Actually correcly annoate mypy stub for image
